requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
pycparser==2.22
sqlparse==0.5.3

djangorestframework~=3.15.2